+++ TestScenario/SP003Test.cs
@@ -31,16 +31,19 @@ public class SP003Test {
   public void sP003() {
     driver.Navigate().GoToUrl("http://localhost:5173/user-home/productDetail/id?id=VL10K1");
     driver.Manage().Window.Size = new System.Drawing.Size(1536, 824);
-    driver.FindElement(By.CssSelector(".thumbnail:nth-child(3)")).Click();
         Thread.Sleep(6000);
+        driver.FindElement(By.CssSelector(".thumbnail:nth-child(3)")).Click();
+        Thread.Sleep(3000);
     driver.FindElement(By.CssSelector(".thumbnail:nth-child(4)")).Click();
-        Thread.Sleep(6000);
+        Thread.Sleep(3000);
 
         driver.FindElement(By.CssSelector(".tab-button:nth-child(2)")).Click();
-        Thread.Sleep(6000);
+        Thread.Sleep(3000);
 
         js.ExecuteScript("window.scrollTo(0,332)");
-    driver.FindElement(By.CssSelector(".tab-button:nth-child(1)")).Click();
-    driver.FindElement(By.CssSelector(".thumbnail:nth-child(5)")).Click();
+        Thread.Sleep(3000);
+        driver.FindElement(By.CssSelector(".tab-button:nth-child(1)")).Click();
+        Thread.Sleep(3000);
+        driver.FindElement(By.CssSelector(".thumbnail:nth-child(5)")).Click();
   }
 }
