+++ TestScenario/SP05Test.cs
@@ -31,14 +31,23 @@ public class SP05Test {
   public void sP05() {
     driver.Navigate().GoToUrl("http://localhost:5173/user-home");
     driver.Manage().Window.Size = new System.Drawing.Size(1536, 824);
-    driver.FindElement(By.CssSelector(".product-list:nth-child(2) .product:nth-child(1) .product-image")).Click();
-    js.ExecuteScript("window.scrollTo(0,0)");
-    driver.FindElement(By.CssSelector(".product-card:nth-child(2) .product-image")).Click();
-    driver.FindElement(By.CssSelector(".add-to-cart > button:nth-child(3)")).Click();
-    driver.FindElement(By.CssSelector(".add-to-cart > button:nth-child(3)")).Click();
-    driver.FindElement(By.CssSelector(".add-to-cart > button:nth-child(1)")).Click();
-    driver.FindElement(By.CssSelector(".add-to-cart > button:nth-child(1)")).Click();
-    driver.FindElement(By.CssSelector("button:nth-child(3)")).Click();
-    driver.FindElement(By.CssSelector("button:nth-child(3)")).Click();
+        Thread.Sleep(6000);
+        driver.FindElement(By.CssSelector(".product-list:nth-child(2) .product:nth-child(1) .product-image")).Click();
+        Thread.Sleep(6000);
+        js.ExecuteScript("window.scrollTo(0,0)");
+        Thread.Sleep(6000);
+        driver.FindElement(By.CssSelector(".product-card:nth-child(2) .product-image")).Click();
+        Thread.Sleep(6000);
+        driver.FindElement(By.CssSelector(".add-to-cart > button:nth-child(3)")).Click();
+        Thread.Sleep(6000);
+        driver.FindElement(By.CssSelector(".add-to-cart > button:nth-child(3)")).Click();
+        Thread.Sleep(6000);
+        driver.FindElement(By.CssSelector(".add-to-cart > button:nth-child(1)")).Click();
+        Thread.Sleep(6000);
+        driver.FindElement(By.CssSelector(".add-to-cart > button:nth-child(1)")).Click();
+        Thread.Sleep(6000);
+        driver.FindElement(By.CssSelector("button:nth-child(3)")).Click();
+        Thread.Sleep(6000);
+        driver.FindElement(By.CssSelector("button:nth-child(3)")).Click();
   }
 }
