+++ TestScenario/SP08Test.cs
@@ -12,7 +12,7 @@ using OpenQA.Selenium.Support.UI;
 using OpenQA.Selenium.Interactions;
 using NUnit.Framework;
 [TestFixture]
-public class SP08Test {
+public class SP09Test {
   private IWebDriver driver;
   public IDictionary<string, object> vars {get; private set;}
   private IJavaScriptExecutor js;
@@ -28,24 +28,18 @@ public class SP08Test {
         driver.Dispose();
     }
   [Test]
-  public void sP08() {
-    driver.Navigate().GoToUrl("http://localhost:5173/user-home");
-    driver.Manage().Window.Size = new System.Drawing.Size(1052, 806);
-    {
-      var element = driver.FindElement(By.CssSelector(".icon-wrapper > a > img"));
-      Actions builder = new Actions(driver);
-      builder.MoveToElement(element).Perform();
+  public void sP09() {
+    driver.Navigate().GoToUrl("http://localhost:5173/user-home/productDetail/id?id=N14K1");
+        Thread.Sleep(6000);
+        driver.Manage().Window.Size = new System.Drawing.Size(1536, 824);
+        Thread.Sleep(6000);
+        driver.FindElement(By.CssSelector(".tab-button:nth-child(2)")).Click();
+        Thread.Sleep(6000);
+        driver.FindElement(By.CssSelector("textarea")).Click();
+        Thread.Sleep(6000);
+        driver.FindElement(By.CssSelector(".comment-input-container > button")).Click();
+        Thread.Sleep(6000);
+        Assert.That(driver.SwitchTo().Alert().Text, Is.EqualTo("Vui lòng nhập nội dung bình luận!"));
+        Thread.Sleep(6000);
     }
-    driver.FindElement(By.CssSelector(".product-list:nth-child(2) .product:nth-child(1) .product-image")).Click();
-    js.ExecuteScript("window.scrollTo(0,0)");
-    driver.FindElement(By.CssSelector(".product-card:nth-child(1) .product-image")).Click();
-    driver.FindElement(By.CssSelector(".tab-button:nth-child(2)")).Click();
-    js.ExecuteScript("window.scrollTo(0,316)");
-    driver.FindElement(By.CssSelector("textarea")).Click();
-    driver.FindElement(By.CssSelector(".product-detail")).Click();
-    driver.FindElement(By.CssSelector(".comment-input-container > button")).Click();
-    Assert.That(driver.SwitchTo().Alert().Text, Is.EqualTo("Vui lòng nhập nội dung bình luận!"));
-    driver.FindElement(By.CssSelector("textarea")).Click();
-    driver.FindElement(By.CssSelector(".comment-input-container > button")).Click();
-  }
 }
